Move public api into main.py
This excludes "serve" function which is in server.py
