+++ jsonrpcserver/__init__.py
@@ -1,5 +1,5 @@
 # from .async_dispatcher import dispatch as async_dispatch
-from .dispatcher import dispatch
+from .main import dispatch, dispatch_to_serializable, dispatch_to_response
 from .methods import add as method
-from .result import Result, SuccessResult, ErrorResult, InvalidParamsResult
+from .result import Result, Success, Error, InvalidParams
 from .server import serve
