+++ jsonrpcserver/server.py
@@ -1,31 +1,20 @@
 import logging
 from http.server import BaseHTTPRequestHandler, HTTPServer
 
-from .dispatcher import dispatch
-
-
-class RequestHandler(BaseHTTPRequestHandler):
-    def do_POST(self) -> None:
-        """HTTP POST"""
-        # Process request
-        request = self.rfile.read(int(str(self.headers["Content-Length"]))).decode()
-        response = dispatch(request)
-        # Return response
-        if response is not None:
-            self.send_response(200)
-            self.send_header("Content-type", "application/json")
-            self.end_headers()
-            self.wfile.write(str(response).encode())
+from .main import dispatch
 
 
 def serve(name: str = "", port: int = 5000) -> None:
-    """
-    A basic way to serve the methods.
+    class RequestHandler(BaseHTTPRequestHandler):
+        def do_POST(self) -> None:
+            response = dispatch(
+                self.rfile.read(int(str(self.headers["Content-Length"]))).decode()
+            )
+            if response is not None:
+                self.send_response(200)
+                self.send_header("Content-type", "application/json")
+                self.end_headers()
+                self.wfile.write(str(response).encode())
 
-    Args:
-        name: Server address.
-        port: Server port.
-    """
     logging.info(" * Listening on port %s", port)
-    httpd = HTTPServer((name, port), RequestHandler)
-    httpd.serve_forever()
+    HTTPServer((name, port), RequestHandler).serve_forever()
